Only perform one trial when determining if a packet is dropped
The previous implementation called error_mode4 twice to determine whether
a packet drop was due to interference or propogation effects. This meant
that two independent trials were performed on the packet drop as
error_mode4 does remember state between calls. The two independent trials
leads to overestimating the packet drop rates as the probability of a
transmission being correctly received with two trials is
           P_success = (1 - (P_tb_err)^2) (1 - (P_sci_err)^2)
rather than
                P_success = (1 - P_tb_err) (1 - P_sci_err)
in the one trial case. This was clearly seen in the offset between the
expected BLER (given by the BLER data being used) and the simulated
BLER. When compared the simulated BLER followed the two trial curves,
rather than the raw BLER curve, provided by the dataset.

This patch corrects this by changing error_mode4 function to perform
both calculations using one random value. This means only one trial is
performed but the desired granularity, being able to determine if a
packet was dropped because of interference of propogation effects, is
maintained. Instead of being called twice, error_mode4 returns a
tuple of values indicating whether the packet was received correctly
based on the interference and the propogation effects. Now, the BLER
follows the expected value for the appropriate BLER data set.