Frequency fix (#7)
* Changes to MAC layer to ensure correct setting of resourceReselectionCounter

* Change fixes issue where sub 100ms transmission rates don't generate grants correctly

* Changes ensure attenuation shadowing is allowed in analytical runs

* WIP: Latest changes to model for validatioN

* Fix issue with snr calc

* Additional statistics

* Error calculation based on the TB not the RB

* Update stat recording

* Fix ensures SINR calc is based on used RBs not all

* Update attenuation returned

* Revert to no shadowed attenuation

* Update to PHY layer to ensure if period is less than 100 the max latency is reduced to it

* Update allows for RSRP based filtering as well as no filtering of possible CSRs

* Update ensures the number of granted blocks is kept consistent for adjacent and non-adjacent modes

* Just ensures consistency for the sci message in terms of the granted blocks

* Addition of a random scheduling mode for Mode 4

* Update allows for non-periodic traffic to be marked as single use transmissions

* Update to fix fading for nakagami, WIP need to confirm working correctly

* Update allows for single slot transmissions and adds some additional result recording

* Change ensure fading implemented correctly and removes issue with failure due to delete not taking place

* Update to 3GPP scenarios ensuring correct configuration

* Update allows for sensing window to be manually changed without changing pstep which affect calculations

* Change ensures position is logged at 1ms intervals

* Update ensures SCI and TB transmissions spread across all rbs allocated

* Only perform one trial when determining if a packet is dropped

The previous implementation called error_mode4 twice to determine whether
a packet drop was due to interference or propogation effects. This meant
that two independent trials were performed on the packet drop as
error_mode4 does remember state between calls. The two independent trials
leads to overestimating the packet drop rates as the probability of a
transmission being correctly received with two trials is
           P_success = (1 - (P_tb_err)^2) (1 - (P_sci_err)^2)
rather than
                P_success = (1 - P_tb_err) (1 - P_sci_err)
in the one trial case. This was clearly seen in the offset between the
expected BLER (given by the BLER data being used) and the simulated
BLER. When compared the simulated BLER followed the two trial curves,
rather than the raw BLER curve, provided by the dataset.

This patch corrects this by changing error_mode4 function to perform
both calculations using one random value. This means only one trial is
performed but the desired granularity, being able to determine if a
packet was dropped because of interference of propogation effects, is
maintained. Instead of being called twice, error_mode4 returns a
tuple of values indicating whether the packet was received correctly
based on the interference and the propogation effects. Now, the BLER
follows the expected value for the appropriate BLER data set.

* Use the BLER calculated with the SINR when checking if a packet dropped due to interference

* Mode4 Validation scnearios updated

* Mode 4 folder updated with latest simulation scenarios

* Remove code which is for my own use as opposed to others

Co-authored-by: Samuel Pell <sam.pell@pg.canterbury.ac.nz>